Modificato tipo variabile pendenza

+++ src/main/java/com/agriflux/agrifluxshared/dto/MorfologiaDTO.java
@@ -6,7 +6,7 @@ public class MorfologiaDTO {
 	
 	private long idMorfologia;
 	private BigDecimal estensioneTerreno;
-	private int pendenza;
+	private BigDecimal pendenza;
 	private String esposizione;
 	private String litologia;
 	
@@ -26,11 +26,11 @@ public class MorfologiaDTO {
 		this.estensioneTerreno = estensioneTerreno;
 	}
 	
-	public int getPendenza() {
+	public BigDecimal getPendenza() {
 		return pendenza;
 	}
 	
-	public void setPendenza(int pendenza) {
+	public void setPendenza(BigDecimal pendenza) {
 		this.pendenza = pendenza;
 	}
